Allow setting late margin on register expirations
When an endpoint registers to us we internally mark the expiration as
some seconds longer than the actual registration.  Previously this
value was fixed at 60 seconds.

Some people need this value to be shorter so they can meet their SLA
by taking a different action when a device doesn't re-register when
expected.

This commit adds a SIP profile parameter sip-expires-late-margin which
allows setting the margin value we apply here.

FS-6101 --resolve

Thanks-to: Emmanuel Schmidbauer <e.schmidbauer@gmail.com>