Require libedit-dev be available unless disabled
Previously we enabled libedit support opportunistically if it was
neither enabled nor disabled explicitly.  But this can result in
people being confused by libedit features not being available when
they really just need to install dependencies.  With this change,
we'll break on configure if libedit-dev isn't there.  If the user
really wants to build without libedit they just need to explicitly
disable it.

Cc: Ken Rice <krice@freeswitch.org>