Revert addition of PERL_GCC_PEDANTIC to CFLAGS
Brian and I decided to handle this by just stripping -ansi -pedantic
from the mod_perl build instead.

Revert "Since we can't tell if the system perl was built with
pedantic, we have a problem because we default debug enabled which
uses AX_CFLAGS_WARN_ALL_ANSI, For the GNU CC compiler it will be
-Wall (and -ansi -pedantic) The result is added to the shellvar being
CFLAGS by default.  Which then gets included for building mod_perl,
Either way I think this is the correct corse of action to overcome
these new builds of mod_perl."

This reverts commit cb94340e268466d19f41961f3b435897994ec1fa.