mod_sofia: Set sip_to_tag on ringing indication for inbound channels.
When bridging a call, the to-tag used in the outgoing 180 Ringing
message for the inbound channel is unavailable until the channel has
been answered. For the outgoing channel this value is already available
through the sip_to_tag variable via the event socket.

This is solved this by setting sip_to_tag to the local leg's tag when
receiving a ringing indication for inbound channels. This will also make
the variable available in the CHANNEL_PROGRESS event through event
socket.

FS-7137 #resolve