FS-7799 Add API command uuid_write_png to mod_png
This adds the API command uuid_write_png to mod_png which takes a uuid and
file as input. This command takes a screenshot from the session associated
with the given uuid and writes it to the given file. If an absolute path
is not specified, the images directory is used as a prefix.