FS-9683: Pass call recovery status to Verto
Previously, users implementing a Verto workflow where both user login and
placing a new call are automated (eg., visiting a URL, and the videoconference
loads automatically) faced the challenge of not having a reliable way to know
that a page load will result in Verto's call recovery mechanism reconnecting
an existing detached call or not.

This adds a verto.clientReady JSON-RPC message, emitted after all calls are
re-attached, which contains a 'reattached_sessions' array, containing the
sess_id of any reattached session.

Client side, this can be caught in the onMessage handler, under the
$.verto.enum.message.clientReady key.